+++ Source/FileBasedApp/FileBasedApp.Toolkit.Tests/AbsoluteWebUriHttpExtensionsTest.cs
@@ -1,0 +1,20 @@
+﻿using System.Net.Http;
+using FileBasedApp.Toolkit;
+using FluentAssertions;
+using JetBrains.Annotations;
+using Xunit;
+
+namespace FileBasedApp.Toolkit.Tests;
+
+[TestSubject(typeof(AbsoluteWebUriHttpExtensions))]
+public class AbsoluteWebUriHttpExtensionsTest
+{
+    [Fact]
+    public void WithBaseAddress_CorrectlySetsAddress()
+    {
+        var client = new HttpClient();
+        var url = AbsoluteWebUri.Create("https://some.url");
+        client = client.WithBaseAddress(url);
+        client.BaseAddress.Should().Be(url.Uri);
+    }
+}